test-requirements.txt
# TO-DO: add python packages that are required for testing this collection
pytest-ansible
pytest-xdist
molecule
tox-ansible
